+++ src/hamplicons/main.py
@@ -1,23 +1,24 @@
 from collections import defaultdict, namedtuple
-import json
-import logging
-import multiprocessing
-import os
-import shutil
-import sys
-import datetime
+from json import dumps as json_dumps
+from multiprocessing import cpu_count, Pool
+from shutil import which
+from datetime import datetime
 from pathlib import Path
 from argparse import ArgumentParser, ArgumentDefaultsHelpFormatter
 from gzip import open as gzopen
 from itertools import combinations
 from subprocess import Popen
-
+import logging
+import os
+import sys
+import pandas as pd
+from openpyxl import Workbook
+from openpyxl.utils.dataframe import dataframe_to_rows
 import coloredlogs
-import xlsxwriter
 
 __version__ = "4.2.0"
 
-## Input constants
+# Input constants
 # Expected format/extension of input files
 READ_FORMAT = ".fastq.gz"
 # Minimum read length, used to exclude primer dimers, etc.
@@ -30,27 +31,24 @@ TARGET_LOCAL_THRESHOLD_READS = 10
 # minimum percentage of reads observed with a specific indel to be reported
 TARGET_LOCAL_THRESHOLD_PERCENTAGE = 0.05
 
-## Unmatched reporting constants
+# Unmatched reporting constants
 # report significant products with >= this reads
 SIGNIFICANT_UNMATCHED_THRESHOLD_READS = 500
 # require these products to be at least this long, to exclude probable primer-dimers
 SIGNIFICANT_UNMATCHED_THRESHOLD_LENGTH = 50
 
-## Barcode matching constants
+# Barcode matching constants
 # bp in start/end that must match to categorize fastq reads as a specific amplicon
 BARSIZE = 30
 # when performing additional matching
 REDUCED_BAR = 15
 
-## Other constants
+# Other constants
 # Size of regions compared between PCR products to avoid too similar sequences
 WORRY_ZONE = 40
 
 
-TargetSequence = namedtuple(
-    "TargetSequence",
-    ["name", "seq", "len", "barcode_5p", "barcode_3p", "ham_5p", "ham_3p"],
-)
+TargetSequence = namedtuple("TargetSequence", ["name", "seq", "len", "barcode_5p", "barcode_3p", "ham_5p", "ham_3p"],)
 
 
 def run_command(command, **kwargs):
@@ -63,18 +61,9 @@ def setup_logging(level=logging.INFO):
     if not sys.stdout.isatty():
         fmt = "%(asctime)s - %(levelname)s - %(message)s"
 
-    coloredlogs.install(
-        level=level,
-        fmt=fmt,
-        datefmt="%H:%M:%S",
-        level_styles={
-            "info": {"color": "white"},
-            "critical": {"color": "red", "bold": True},
-            "verbose": {"color": "blue"},
-            "error": {"color": "red"},
-            "debug": {"color": "green"},
-            "warning": {"color": "yellow"},
-        },
+    coloredlogs.install(level=level, fmt=fmt, datefmt="%H:%M:%S", level_styles={
+        "info": {"color": "white"}, "critical": {"color": "red", "bold": True}, "verbose": {"color": "blue"},
+        "error": {"color": "red"}, "debug": {"color": "green"}, "warning": {"color": "yellow"},},
     )
 
     return logging.getLogger("main")
@@ -103,23 +92,23 @@ def hamming(seq1, seq2):
 
 
 def run_flash(args):
-    timestamp = datetime.datetime.now()
+    timestamp = datetime.now()
     destination, sample = args
 
     input_1, input_2 = sample["R1"], sample["R2"]
-    output = os.path.join(destination, "index%03i" % (sample["#"],))
-    logfile = output + ".stdout"
+    prefix = "index%03i" % (sample["#"],)
+    logfile = destination / f"{prefix}.stdout"
 
-    command = ["flash", "-t", "1", "-M", "151", "-z", "-o", output, input_1, input_2]
+    command = ["flash", "-t", "1", "-M", "151", "-z", "-o", prefix, "-d", destination, input_1, input_2]
 
     log = logging.getLogger("flash")
-    log.debug("Sample %03i: Merging reads into %s.*", sample["#"], output)
+    log.debug("Sample %03i: Merging reads into %s.*", sample["#"], destination / prefix)
     with open(logfile, "w") as loghandle:
         if run_command(command, stdout=loghandle):
             log.error("error while running FLASH; see log at %r", logfile)
             return False
 
-    delta = (datetime.datetime.now() - timestamp).total_seconds()
+    delta = (datetime.now() - timestamp).total_seconds()
     log.info("Sample %03i merged in %.1f seconds", sample["#"], delta)
 
     return True
@@ -130,7 +119,7 @@ def merge_fastq(destination, samples, threads):
     log.info("Merging FASTQ files")
 
     args = [(destination, sample) for sample in samples]
-    pool = multiprocessing.Pool(processes=threads)
+    pool = Pool(processes=threads)
     for result in pool.imap(run_flash, args):
         if not result:
             pool.terminate()
@@ -153,7 +142,7 @@ def count_fastq_sequences(filename, min_length=MIN_READ_LENGTH):
 
 
 def collect_fastq_files(log, root, recursive):
-    log.info("Locating FASTQ files in %s", root)
+    log.info(f"Locating FASTQ files in {root}")
 
     fastq_files = defaultdict(dict)
     for (dirpath, _, filenames) in os.walk(root, followlinks=True):
@@ -229,9 +218,7 @@ def merge_identical_targets(targets):
     for group in targets_by_barcodes.values():
         if len(group) > 1:
             if any(group[0].seq != target.seq for target in group[1:]):
-                log.error(
-                    "Found targets with identical barcodes but different sequences:"
-                )
+                log.error("Found targets with identical barcodes but different sequences:")
                 for target in group:
                     log.error("    - %s", target.name)
                 return None
@@ -290,14 +277,12 @@ def check_targets(targets, max_ham):
     targets_by_name = {}
     for target in targets:
         if target.name in targets_by_name:
-            log.error(
-                "Multiple, different target sequences with the name %r", target.name
-            )
+            log.error("Multiple, different target sequences with the name %r", target.name)
             return
 
         targets_by_name[target.name] = target
 
-    # 2: itereate each barcode through worry zone
+    # 2: iterate each barcode through worry zone
     for key1, key2 in combinations(targets_by_name, 2):
         target1 = targets_by_name[key1]
         target2 = targets_by_name[key2]
@@ -313,27 +298,15 @@ def check_targets(targets, max_ham):
         ham_3p = hamming(target1.barcode_3p, target2.barcode_3p)
 
         # set nearest ham
-        targets_by_name[target1.name] = target1._replace(
-            ham_5p=min(target1.ham_5p, ham_5p), ham_3p=min(target1.ham_3p, ham_3p)
-        )
-        targets_by_name[target2.name] = target2._replace(
-            ham_5p=min(target2.ham_5p, ham_5p), ham_3p=min(target2.ham_3p, ham_3p)
-        )
+        targets_by_name[target1.name] = target1._replace(ham_5p=min(target1.ham_5p, ham_5p), ham_3p=min(target1.ham_3p, ham_3p))
+        targets_by_name[target2.name] = target2._replace(ham_5p=min(target2.ham_5p, ham_5p), ham_3p=min(target2.ham_3p, ham_3p))
 
     targets = sorted(targets_by_name.values(), key=lambda item: item.name.lower())
 
     name_len = max(len(x.name) for x in targets)
-    log.info(
-        "{:<{}} {:>8} {:>8} {:>8}".format(
-            "Name", name_len, "Length", "5p Ham", "3p Ham"
-        )
-    )
+    log.info("{:<{}} {:>8} {:>8} {:>8}".format("Name", name_len, "Length", "5p Ham", "3p Ham"))
     for target in targets:
-        log.info(
-            "{:<{}} {:>8} {:>8} {:>8}".format(
-                target.name, name_len, target.len, target.ham_5p, target.ham_3p
-            )
-        )
+        log.info("{:<{}} {:>8} {:>8} {:>8}".format(target.name, name_len, target.len, target.ham_5p, target.ham_3p))
 
     return targets
 
@@ -342,18 +315,10 @@ def build_targets(fasta, max_ham):
     targets = []
     for (name, seq) in fasta:
         seq = seq.upper()
-
-        targets.append(
-            TargetSequence(
-                name=name,
-                seq=seq,
-                len=len(seq),
-                barcode_5p=seq[:BARSIZE],
-                barcode_3p=seq[-BARSIZE:],
-                ham_5p=max_ham,
-                ham_3p=max_ham,
-            )
-        )
+        targets.append(TargetSequence(
+            name=name, seq=seq, len=len(seq), barcode_5p=seq[:BARSIZE], barcode_3p=seq[-BARSIZE:],
+            ham_5p=max_ham, ham_3p=max_ham,
+        ))
 
     return check_targets(targets, max_ham)
 
@@ -475,9 +440,7 @@ def analyze_sample(parameters):
     for target in targets:
         results[target.name] = defaultdict(int)
 
-    filename = os.path.join(
-        args.output_merged, "index%03i.extendedFrags.fastq.gz" % (sample["#"],)
-    )
+    filename = os.path.join(args.output_merged, "index%03i.extendedFrags.fastq.gz" % (sample["#"],))
     log.debug("Sample %03i: Reading index from %s", sample["#"], filename)
     read_counts = count_fastq_sequences(filename)
     total_reads = sum(read_counts.values())
@@ -488,46 +451,23 @@ def analyze_sample(parameters):
         logstats[target.name] = [0, 0, 0, 0, 0, 0]
 
     # STRICT matching is very fast and assigns anything we can easily identify
-    log.debug(
-        "Sample %03i: Strict matching %i products with %i reads",
-        sample["#"],
-        len(read_counts),
-        sum(read_counts.values()),
-    )
+    log.debug("Sample %03i: Strict matching %i products with %i reads", sample["#"], len(read_counts), sum(read_counts.values()),)
     strict_matching(targets, read_counts, results, logstats)
 
     if not args.deactivate_relaxed_matching:
         # RELAXED matching (slower, but not very)
-        log.debug(
-            "Sample %03i: Relaxed matching %i products with %i reads",
-            sample["#"],
-            len(read_counts),
-            sum(read_counts.values()),
-        )
+        log.debug("Sample %03i: Relaxed matching %i products with %i reads", sample["#"], len(read_counts), sum(read_counts.values()),)
         relaxed_matching(args, targets, read_counts, results, logstats)
 
         # ULTRA RELAXED MATCHING (probably slow)
         # now anything left will be ultra relaxed matched. This opens up for a read to
-        # match multiple targets so we need to check for this and also any offtargets
+        # match multiple targets, so we need to check for this and also any off-targets
         # these primers may be producing will also show up here...maybe that should be
         # a number to report...
-        log.debug(
-            "Sample %03i: Ultra relaxed matching %i products with %i reads",
-            sample["#"],
-            len(read_counts),
-            sum(read_counts.values()),
-        )
+        log.debug("Sample %03i: Ultra relaxed matching %i products with %i reads", sample["#"], len(read_counts), sum(read_counts.values()),)
         ultra_relaxed_matching(args, targets, read_counts, results, logstats)
 
-    return {
-        "#": sample["#"],
-        "name": sample["name"],
-        "#reads": total_reads,
-        "#unmatched": sum(read_counts.values()),
-        "targets": results,
-        "logstats": logstats,
-        "unmatched": read_counts,
-    }
+    return {"#": sample["#"], "name": sample["name"], "#reads": total_reads, "#unmatched": sum(read_counts.values()), "targets": results, "logstats": logstats, "unmatched": read_counts}
 
 
 def log_sample_stats(sample):
@@ -567,73 +507,53 @@ def analyze(args, samples, fasta):
     log = logging.getLogger("matching")
 
     results = []  # results are [{name: defaultdict(int), ...}, ...]
-    pool = multiprocessing.Pool(processes=args.threads)
+    pool = Pool(processes=args.t)
     pool_args = [(args, sample, fasta) for sample in samples]
     for output in pool.imap(analyze_sample, pool_args):
-        log.info("Completed analysis of sample %i ..", output["#"])
+        log.info(f"Completed analysis of sample {output['#']} ..")
         log_sample_stats(sample=output)
 
         output.pop("unmatched")
         results.append(output)
 
     if results:
-        output_summary = {
-            "#": "*",
-            "name": "summary",
-            "logstats": results[0].pop("logstats"),
-            "unmatched": {},
-        }
+        output_summary = {"#": "*", "name": "summary", "logstats": results[0].pop("logstats"), "unmatched": {}, }
 
         for result in results[1:]:
             for name, numbers in result.pop("logstats").items():
                 for idx, count in enumerate(numbers):
                     output_summary["logstats"][name][idx] += count
 
-        log.info("Completed analysis of %i samples!", len(results))
+        log.info(f"Completed analysis of {len(results)} samples!")
         log_sample_stats(sample=output_summary)
 
     return results
 
 
 def write_report(args, results, output_filename):
-    workbook = xlsxwriter.Workbook(output_filename)
-    thousandsepstyle = workbook.add_format({"num_format": "#,##0"})
-    percentstyle = workbook.add_format({"num_format": "0.00%"})
+    workbook = Workbook()
 
-    worksheet = workbook.add_worksheet("results")
-    worksheet.set_column(2, 2, 20)
-    worksheet.write(0, 0, "Index/Well")
-    worksheet.write(0, 1, "Sample")
-    worksheet.write(0, 2, "Target")
-    worksheet.write(0, 3, "Reads")
-    worksheet.write(0, 4, "WT")
-    worksheet.write(0, 5, "Indel")
-    worksheet.write(0, 6, "%Indel")
+    worksheet1 = workbook.active
+    worksheet1.title = "results"
+    worksheet1.cell(1, 1, "Index/Well")
+    worksheet1.cell(1, 2, "Sample")
+    worksheet1.cell(1, 3, "Target")
+    worksheet1.cell(1, 4, "Reads")
+    worksheet1.cell(1, 5, "WT")
+    worksheet1.cell(1, 6, "Indel")
+    worksheet1.cell(1, 7, "%Indel")
     # peak counting
     max_peaks = 0
     for sample in results:
         for target, peaks in sample["targets"].items():
-            sigpeaks = len(
-                [
-                    p
-                    for s, p in peaks.items()
-                    if s
-                    and p >= TARGET_LOCAL_THRESHOLD_READS
-                    and (float(p) / sum(peaks.values()))
-                    >= TARGET_LOCAL_THRESHOLD_PERCENTAGE
-                ]
-            )
+            sigpeaks = len([p for s, p in peaks.items() if s and p >= TARGET_LOCAL_THRESHOLD_READS and (float(p) / sum(peaks.values())) >= TARGET_LOCAL_THRESHOLD_PERCENTAGE])
             if sigpeaks > max_peaks:
                 max_peaks = sigpeaks
     # add a header per peak
     peak_col_start = 7
-    if max_peaks:
-        worksheet.set_column(peak_col_start, max_peaks + peak_col_start - 1, 16)
     for col in range(peak_col_start, max_peaks + peak_col_start):
-        worksheet.write(0, col, "peak{}".format(col - (peak_col_start - 1)))
-        worksheet.write(
-            0, col + max_peaks, "peak{}%".format(col - (peak_col_start - 1))
-        )
+        worksheet1.cell(1, 1+col, "peak{}".format(col - (peak_col_start - 1)))
+        worksheet1.cell(1, 1+col + max_peaks, "peak{}%".format(col - (peak_col_start - 1)))
     row = 1
     for sample in results:
         for target, peaks in sample["targets"].items():
@@ -642,17 +562,15 @@ def write_report(args, results, output_filename):
                     rows, columns = args.plate_layout
                     well = convert_index_to_well(sample["#"], rows, columns)
 
-                    worksheet.write(row, 0, well)
+                    worksheet1.cell(row+1, 1, well)
                 else:
-                    worksheet.write(row, 0, sample["#"])
-                worksheet.write(row, 1, sample["name"])
-                worksheet.write(row, 2, target)
-                worksheet.write(row, 3, sum(peaks.values()), thousandsepstyle)
-                worksheet.write(row, 4, peaks[0], thousandsepstyle)
-                worksheet.write(
-                    row, 5, sum(peaks.values()) - peaks[0], thousandsepstyle
-                )
-                worksheet.write(row, 6, "=F{0}/D{0}".format(row + 1), percentstyle)
+                    worksheet1.cell(row+1, 1, sample["#"])
+                worksheet1.cell(row+1, 2, sample["name"])
+                worksheet1.cell(row+1, 3, target)
+                worksheet1.cell(row+1, 4, sum(peaks.values())).style = 'Comma [0]'
+                worksheet1.cell(row+1, 5, peaks[0]).style = "Comma [0]"
+                worksheet1.cell(row+1, 6, sum(peaks.values()) - peaks[0]).style = "Comma [0]"
+                worksheet1.cell(row+1, 7, "=F{0}/D{0}".format(row + 1)).style = 'Percent'
                 column = 7
                 for indel_size, num in sorted(peaks.items()):
                     if indel_size:
@@ -663,20 +581,26 @@ def write_report(args, results, output_filename):
                         ):
                             if indel_size % 3 == 0:
                                 indel_size = "{} (inframe)".format(indel_size)
-                            worksheet.write(row, column, "{}".format(indel_size))
-                            worksheet.write_formula(
-                                row,
-                                column + max_peaks,
-                                "={0}/D{1}".format(num, row + 1),
-                                percentstyle,
-                            )
+                            worksheet1.cell(row+1, column+1, "{}".format(indel_size))
                             column += 1
+                            worksheet1.cell(row+1, column + max_peaks, f"={num}/D{row + 1}").style = 'Percent'
                 row += 1
-    worksheet.autofilter(
-        0, 0, row, (max_peaks * 2) + (peak_col_start - 1)
-    )  # Same as above.
+    worksheet1.auto_filter.ref = worksheet1.dimensions
 
-    workbook.close()
+    for cs in worksheet1.columns:
+        worksheet1.column_dimensions[cs[0].column_letter].bestFit = True
+
+    indel_sizes = []
+    for i in results:
+        q = pd.DataFrame(i['targets']).T
+        q['index'] = i['#']
+        indel_sizes.append(q)
+    indel_df = pd.concat(indel_sizes).set_index('index', append=True, drop=True).rename_axis(('target', 'index')).sort_index(axis=1).reset_index()
+    worksheet2 = workbook.create_sheet("indel_sizes")
+    for r in dataframe_to_rows(indel_df, index=False, header=True):
+        worksheet2.append(r)
+    worksheet2.auto_filter.ref = worksheet2.dimensions
+    workbook.save(output_filename)
 
 
 def write_json(args, targets, results, output_filename):
@@ -716,7 +640,7 @@ def write_json(args, targets, results, output_filename):
         }
 
     with open(output_filename, "wt") as handle:
-        handle.write(json.dumps(data))
+        handle.write(json_dumps(data))
 
 
 def parse_args():
@@ -726,100 +650,25 @@ def parse_args():
         formatter_class=ArgumentDefaultsHelpFormatter,
     )
 
-    parser.add_argument(
-        "output_prefix", nargs="?", default="output", help="Prefix for output files"
-    )
-
-    parser.add_argument(
-        "--version", action="version", version="%(prog)s v" + __version__
-    )
-    parser.add_argument(
-        "--threads",
-        type=int,
-        help="Number of threads for read merging/analysis",
-        default=multiprocessing.cpu_count(),
-        metavar="N",
-    )
-    parser.add_argument(
-        "--log-level",
-        type=str.upper,
-        default="INFO",
-        choices=("DEBUG", "INFO", "WARNING", "ERROR"),
-        help="Log-level of messages printed STDERR and written to ${output_prefix}.log",
-    )
+    parser.add_argument('-v', action="version", version="%(prog)s v" + __version__)
+    parser.add_argument('-t', type=int, help="Number of threads for read merging/analysis", default=cpu_count(), metavar='N')
+    parser.add_argument("--log-level", type=str.upper, default="INFO", choices=("DEBUG", "INFO", "WARNING", "ERROR"), help="Log-level of messages printed STDERR and written to hamp_out.log")
 
     group = parser.add_argument_group("Input")
-    group.add_argument(
-        "-tf",
-        "--targets-fasta",
-        type=str,
-        default="targets.fa",
-        help="Fasta file containing WT PCR sequences",
-        metavar="TARGETS",
-    )
-    group.add_argument(
-        "-dd",
-        "--data-directory",
-        type=str,
-        default="fastq",
-        help="Directory containing fastq.gz files",
-        metavar="DIRECTORY",
-    )
-    group.add_argument(
-        "-dd-recursive",
-        "--data-directory-recursive",
-        action="store_true",
-        help="Search data directory recursively",
-    )
-    group.add_argument(
-        "-sm",
-        "--skip-merging",
-        action="store_true",
-        help="skip merging of fastq paired end reads",
-    )
+    group.add_argument("-tf", "--targets-fasta", type=str, default="targets.fa", help="Fasta file containing WT PCR sequences", metavar="TARGETS",)
+    group.add_argument("-dd", "--data-directory", type=str, default="fastq", help="Directory containing fastq.gz files", metavar="DIRECTORY",)
+    group.add_argument("-ddr", "--data-directory-recursive", action="store_true", help="Search data directory recursively", )
+    group.add_argument("-sm", "--skip-merging", action="store_true", help="skip merging of fastq paired end reads",)
 
     group = parser.add_argument_group("Matching")
-    group.add_argument(
-        "-dr",
-        "--deactivate-relaxed-matching",
-        action="store_true",
-        help="Deactivate relaxed matching",
-    )
-    group.add_argument(
-        "-hd",
-        "--hamming-distance",
-        type=int,
-        help="Allow this many mismatches in relaxed matching",
-        default=4,
-        metavar="INT",
-    )
+    group.add_argument("-dr", "--deactivate-relaxed-matching", action="store_true", help="Deactivate relaxed matching",)
+    group.add_argument("-hd", "--hamming-distance", type=int, help="Allow this many mismatches in relaxed matching", default=4, metavar="INT",)
 
     group = parser.add_argument_group("Output")
-    group.add_argument(
-        "-of",
-        "--output-formats",
-        nargs="+",
-        choices=("json", "xlsx"),
-        metavar="FORMAT",
-        default="xlsx",
-        help="One or more formats (xlsx or json) in which to save analysis results",
-    )
-    group.add_argument(
-        "-cw",
-        "--convert-to-wells",
-        action="store_true",
-        help="Convert Index to Well",
-    )
-    group.add_argument(
-        "-pl",
-        "--plate-layout",
-        type=list,
-        nargs=2,
-        help="Define layout for index translation when using --convert-to-wells. "
-        "E.g. `-pl 8 12` for 8 rows and 12 columns",
-        default=[8, 12],
-        metavar="INT",
-    )
+    group.add_argument('-o', default="output", help="Folder to put all output in", type=Path, metavar="output_folder")
+    group.add_argument("-of", "--output-formats", nargs="+", choices=("json", "xlsx"), metavar="FORMAT", default="xlsx", help="One or more formats (xlsx or json) in which to save analysis results",)
+    group.add_argument("-cw", "--convert-to-wells", action="store_true", help="Convert Index to Well", )
+    group.add_argument("-pl", "--plate-layout", type=list, nargs=2, help="Define layout for index translation when using --convert-to-wells. " "E.g. `-pl 8 12` for 8 rows and 12 columns", default=[8, 12], metavar="INT",)
 
     args = parser.parse_args()
     # check existence of input files
@@ -833,59 +682,43 @@ def main(args):
     log = setup_logging(level=args.log_level)
 
     # if we are merging, check that flash exists
-    if not (args.skip_merging or shutil.which("flash")):
+    if not (args.skip_merging or which("flash")):
         log.error("Required executable `flash` not found in PATH!")
         log.error("Please install FLASH from 'http://ccb.jhu.edu/software/FLASH/'")
 
         return 1
 
-    args.output_merged = args.output_prefix + ".merged"
-    os.makedirs(args.output_merged, exist_ok=True)
+    args.output_merged = args.o / "hamp_out.merged"
+    args.output_merged.mkdir(parents=True, exist_ok=True)
 
-    setup_file_logging(filename=args.output_prefix + ".log", level=args.log_level)
+    setup_file_logging(filename=args.o / "hamp_out.log", level=args.log_level)
 
-    log.info("Reading FASTA file %r", args.targets_fasta)
-    targets = build_targets(
-        read_fasta(args.targets_fasta),
-        max_ham=args.hamming_distance,
-    )
+    log.info(f"Reading FASTA file {args.targets_fasta}")
+    targets = build_targets(read_fasta(args.targets_fasta), max_ham=args.hamming_distance)
 
     if targets is None:
         return 1
 
-    samples = collect_fastq_files(
-        log,
-        args.data_directory,
-        recursive=args.data_directory_recursive,
-    )
+    samples = collect_fastq_files(log, args.data_directory,recursive=args.data_directory_recursive,)
 
     if samples is None:
         return 1
 
     if args.skip_merging:
         log.info("Skipping FASTQ merging")
-    elif not merge_fastq(args.output_merged, samples, threads=args.threads):
+    elif not merge_fastq(args.output_merged, samples, args.t):
         return 1
 
     log.info("Analyzing FASTQ files")
     results = analyze(args, samples, targets)
 
     if "xlsx" in args.output_formats:
-        log.info("Writing report to %r", args.output_prefix + ".xlsx")
-        write_report(
-            args=args,
-            results=results,
-            output_filename=args.output_prefix + ".xlsx",
-        )
+        log.info(f"Writing report to {args.o / 'hamp_out.xlsx'}")
+        write_report(args=args, results=results, output_filename=args.o / "hamp_out.xlsx",)
 
     if "json" in args.output_formats:
-        log.info("Writing JSON to %r", args.output_prefix + ".json")
-        write_json(
-            args=args,
-            targets=targets,
-            results=results,
-            output_filename=args.output_prefix + ".json",
-        )
+        log.info(f"Writing JSON to {args.o / 'hamp_out.json'}")
+        write_json(args=args, targets=targets, results=results, output_filename=args.o / "hamp_out.json",)
 
     log.info("Done")